Refactor pcapng: move InterfaceStatistics to its own file
